Create Code Fusing 1
This is the first version of adding sections of code together from the multiple choice code and the gametracker code. This code does a very good job with using the class feature for the questions, answers and prompts parameters and lists. It also makes great use of defining and calling functions. Based on this initial fusion, other sections of code will separately spring from this (separate quiz sections), get corrected and fused into the final version of PocketNurse.