Update Code Fusing 2 Test Shell
Code was added in to remove the "\n" from the output when printing out the questions that the student needs to review. Code was also added in to arrange the review list of questions on new lines. This will present the questions to be reviewed in an easily readable format.